PB: Get rid of Windows typedefs where applicable

+++ Extract/cpz/PB.h
@@ -5,5 +5,5 @@ class CPB
 public:
 	virtual ~CPB() = default;
 
-	void DecompLZSS(u8* dst, DWORD dst_size, const u8* flags, DWORD flags_size, const u8* src, DWORD src_size);
+	void DecompLZSS(u8* dst, size_t dst_size, const u8* flags, size_t flags_size, const u8* src, size_t src_size);
 };
